refactor control-plane detection lifecycle boundary
